functions/kill-duplicate-jobs/requirements.txt
google-cloud-storage==1.9.0